requirements.txt
streamlit==1.30.0
pandas==2.1.3
numpy==1.26.0
scikit-learn==1.3.2
anthropic>=0.18.1
python-dotenv==1.0.0
line-bot-sdk>=3.5.0,<4.0.0
requests==2.31.0
matplotlib==3.8.2
seaborn==0.13.0